add support for whitespace-delimited urls, unlocking ENV/Keystore expansion
Since literal whitespace is illegal in a URL, we can safely split on it
to provide support for several hosts to be specified in a single string.

This empowers users to use Environment- and Keystore-variables to populate
an arbitrary number of hosts.

For example, with the environment:

~~~ sh
export ES_HOSTS="es1.mydomain.com es2.mydomain.com"
~~~

The following configurations are equivalent:
~~~
output {
  hosts => "${ES_HOSTS}"
}
~~~

~~~
output {
  hosts => ["es1.mydomain.com","es2.mydomain.com"]
}
~~~